Specify container format when starting audio decoding.
Add a parameter to SetAudioPid command to specify the container format the audio stream is packetized in. This is important to support DVD PES streams. The PlayAudio function detects now an audio substream header and switches to DVD PES format. Otherwise standard PES format is assumed.

Firmware 0.2.8 is required to support that.