Reset the decoder during device initialization.
It can happen, e. g. after a segfault of the VDR, that during restart the
decoder is still in the same state it was before the crash. Then there
are e. g. still OSD handles open that are not known to the current
instance of the VDR. This could lead to some unwanted effects like an
invisible OSD.

This change introduces some additional calls during device initialization
that reset the decoder to a state similar to the initial one.

Signed-off-by: Andreas Regel <andreas.regel@gmx.de>