libhdffcmd: Add function HdffCmdAvSetSyncShift().
This functions allows setting of a A/V sync shift value that shifts the A/V
sync offset by the specified time in ms. This influences the amount of
buffering that is done in transfer mode before starting to play. Positive
values enlarge the buffer, negative values shrink the buffer.

Signed-off-by: Andreas Regel <andreas.regel@gmx.de>